# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Laravel pour le TP01 - Gestion de location d'équipements sportifs ",
        "description": "Documentation API Laravel avec Swagger",
        "version": "1.0.0"
    },
    "paths": {
        "/api/equipments": {
            "get": {
                "tags": [
                    "Equipment"
                ],
                "summary": "Récupérer la liste des équipements",
                "description": "Retourner tous les équipements existants",
                "operationId": "6ef4946ec56ce0f0c3e613f7c57d37f2",
                "responses": {
                    "200": {
                        "description": "Liste des équipements récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "description": {
                                            "type": "string"
                                        },
                                        "daily_price": {
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "categoryId": {
                                            "type": "integer",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/equipments/{id}": {
            "get": {
                "tags": [
                    "Equipment"
                ],
                "summary": "Récupérer un équipement par ID",
                "description": "Retourner les détails d'un équipement",
                "operationId": "f5219eb4f4143cf304576feea563c147",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Équipement trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "description": {
                                            "type": "string"
                                        },
                                        "daily_price": {
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "categoryId": {
                                            "type": "integer",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Équipement non trouvé"
                    }
                }
            }
        },
        "/api/equipments/{id}/popularity": {
            "get": {
                "tags": [
                    "Equipment"
                ],
                "summary": "Récupérer la popularité d'un équipement par ID",
                "description": "Retourner la popularité d'un équipement",
                "operationId": "5d5c14d5aaa81bf814a18d9d603200ce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Popularité trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "popularity": {
                                            "type": "number",
                                            "format": "float"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Popularité non trouvé"
                    }
                }
            }
        },
        "/api/equipments/{id}/average-rental-price": {
            "get": {
                "tags": [
                    "Equipment"
                ],
                "summary": "Recevoir la moyenne du prix total de location d'un équipement",
                "description": "minDate et maxDate pour filtrer les locations",
                "operationId": "722041b82f3fe0d4c96ce79a786cd2b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "name": "minDate",
                        "in": "query",
                        "description": "AAAA-MM-JJ",
                        "required": false
                    },
                    {
                        "name": "maxDate",
                        "in": "query",
                        "description": "AAAA-MM-JJ",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Moyenne calculée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "average_total_price": {
                                            "type": "number",
                                            "format": "float"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Format de date invalide ou minDate > maxDate"
                    },
                    "404": {
                        "description": "Équipement non trouvé"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Equipment",
            "description": "Equipment"
        }
    ]
}